Allow running sighandler logic as part of a cleanup chain
In our code, we need to take certain cleanup actions
after receiving a SIGSEGV. This makes it difficult to
integrate with backward, as our custom sighandler is overriden.

A simple solution is to have backward-cpp expose the
core sighandling logic as a public function, which
we can simply call from within our code. This is not
currently possible since sig_handler is private, and
additionally raises the signal again at the end.

This pull request makes the above possible. The alternative
is to copy handleSignal into our own code, which is just ugly.