DW_AT_call_file is an unsigned value
According to the spec (http://www.dwarfstd.org/doc/DWARF4.pdf, section 2.14) the DW_AT_call_file attribute (like DW_AT_call_line and DW_AT_call_column) is an Unsigned Integer constant and not a Signed Integer. I was getting crashes in debug symbols created by gcc 4.9.0 for ARM because it was incorrectly getting a very high number that, in backward, was being interpreted as a negative value.