DW_AT_call_file is an unsigned value
According to the DWARF spec, DW_AT_call_file is an unsigned value and we were treating it like a signed integer. Some old gcc compilers for ARM create, probably because of a bug, very high file index numbers which were being interpreted as a negative number, causing a crash.